requirements.txt
# 金价监测脚本依赖
python-dotenv>=1.0.0
